Example of how to fall back to session on load
i.e. If a record isn't found in your database you can build an
ActiveRecord object and the view won't know the difference. As in this
case you can build it from the session if you want to, but most of the
time the session shouldn't have anything useful in it anyway...